fix bug that prevented to combine several time-series with different time-step
the reset for observation t was done at the beginning of the prediction
step leading to time t, which was doomed to occur on the time t’ of the
time-series with smallest time step (t-1 < t’ < t). Now, reset for
observation t is done immediately after the prediction at time t-1,
thus ensuring that prediction at time t doesn’t depend on intermediate
time-step of other time series.